[sh] page5->page2 화면 전환 수정 중

+++ app/src/main/java/com/example/brandol/adaptor/UserCategoryAdapter.kt
@@ -105,7 +105,7 @@ import com.example.brandol.connection.RetrofitClient2
 
 interface ItemClickListener {
     fun onItemClick(position: Int, brandId: Int)
-    fun onItemClick(brandId: Int)
+    fun onItemClick(brandId: Long)
 }
